Customizable hard power off timeout
This is similar to my previous pull request, but:
- branched off of a much more recent commit
- none of the version numbers or firmware builds now that those requirements have changed
- now it enforces a 3-second minimum, if you try to set it lower.
I've tested, and on the first launch with no value set, it correctly uses (and writes) the default of 120 seconds (which matches current behaviour), until you change it to something else in the settings file.
I'm assuming it's desireable for powerOffSecs to not appear in the settings file of hardware that can't use the feature, but if you'd rather the settings be consistent even to the inclusion of unused settings, I can remove those directives wrapping the value inside of keyNames.
FWIW, I've only tested this personally on a TinyTV2.