Settings to resume or shuffle on boot
Two new settings.

If "allowResume" is false, it behaves as previous; every video's playback time starts counting from power-on time.

If "allowResume" is true and "shuffleResume" is false, it writes the current seconds value on power-off and reads the stored value on power-on, so it will come back to (basically) the same place it turned off at.

The arduino string-to-int parsing library seems to want everything to be a 32-bit signed int, if I'm understanding correctly -- but that's still decades of time in seconds, so whatever.  (An overflow appears to just make it start at zero, which is fine for an application where the resume function is just for fun, nothing important depends on it...)

If "allowResume" and "shuffleResume" are both true, the system writes the current millis() value to disk, to be used as a random seed on next boot.

I'm honestly not *entirely* certain how well I'm avoiding noticeable bias in my use of the random functions, but in a dozen or so tests of a video of an hour-long countdown timer, it looks pretty random to me.  I'd welcome any improvements to that portion, though.

The one thing that makes me a little nervous: saving the progress (or the milliseconds seed) based on the moment where the power button was pressed, required making the settingsNeedSaved check shorter in that case, or else the power would be off before it could be performed (I believe this is the case even with a longer hard power timeout time, because the logic that checks for whether a save needs to be flushed, doesn't run after the soft power off regardless?  Though I only half-understand that whole process, actually.)  I had originally made the "write settings on power-off requested" timeout instant, but actually, this caused a noticeable hitch in the static animation, so now it waits long enough for that animation to finish (1 second), which now seems to be fine.   Still, this whole approach does feel SLIGHTLY flimsy;  it might be better for the system to ALWAYS write the settings to disk once the animation is over (rather than waiting on a timer that HAPPENS to last the same length of time), and refuse to hard power down until that writing is complete (rather than enforcing two more seconds in the minimum hard-power-off timeout value in the settings file, which appears to be enough time in my experience, but remains a race condition that's making assumptions about disk speed).   I'm going to propose this as-is for now though, as it seems to work, and a proper overhaul might be better made by someone more comfortable with the project.