Add debounce to TinyTV2.h
I've never worked with rotary encoders before, but a graph in the datasheet suggested that as soon as the pins start changing, their values (including that initiating change) should NOT be trusted for the next several ms until they're guaranteed to have FINISHED bouncing; the value that they STOP at is the actual correct value. So: the pin-watching logic still *stores* whatever the pins say in real-time, BUT, that value does NOT get passed to the rest of the program to be acted on, until the timeout has cleared.

Relatedly: "has the encoder turned CW" and "has the encoder turned CCW" are now stored in the same variable in order to make them mutually exclusive, since otherwise they'd almost inevitably have BOTH wound up true over the course of any bouncing; the point is to keep only the encoder's FINAL answer, ignoring all the other answers that came before (within that single debounce period), so now both possibilities are stored in a single variable so that every new answer will *completely* replace any previous ones.

Full disclosure:  I'm still not totally sure that this is *completely* correct; I have a hunch that maybe the "old pin" values shouldn't be replaced EITHER, until after the debounce period, in order to be strictly correct?  But: in my testing of this change, I've only had two 'bounced' inputs in dozens and dozens of tests, and I'm willing to shrug that off as good enough at this point; but if anybody more used to working with rotary encoders has some insight, I'd welcome it.